addModulsAndSetUpProject
