+++ PomodoroTimer/App/SceneDelegate.swift
@@ -1,0 +1,38 @@
+//
+//  SceneDelegate.swift
+//  PomodoroTimer
+//
+//  Created by Daniil Litvinov on 02.10.2023.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let windosScene = (scene as? UIWindowScene) else { return }
+        self.window = UIWindow(windowScene: windosScene)
+        let navigationController = UINavigationController(rootViewController: MasterScreen())
+        window?.rootViewController = navigationController
+        window?.makeKeyAndVisible()
+    }
+
+    func sceneDidDisconnect(_ scene: UIScene) {
+    }
+
+    func sceneDidBecomeActive(_ scene: UIScene) {
+    }
+
+    func sceneWillResignActive(_ scene: UIScene) {
+    }
+
+    func sceneWillEnterForeground(_ scene: UIScene) {
+    }
+
+    func sceneDidEnterBackground(_ scene: UIScene) {
+    }
+    
+}
+
