+++ PomodoroTimer/Resourse/Info.plist
