--- PomodoroTimer/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  PomodoroTimer
-//
-//  Created by Daniil Litvinov on 02.10.2023.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
